Rock, Paper, Scissors-Supreme
First you are asked for your name. If you enter nothing, it will prompt you again. Afterwards, it explains the rules and asks you for either Schere, Stein or Papier. It picks a random value of the three and compares it to your input. Then, it assigns points to the winner (draw is one point each). If your input is not one of the three values, it will tell you that your input is not valid and that you should try again.